refactor example directory organization Note that this will break some path configurations in the example scripts for the time being.
