+++ example/apps/config_template.yaml
