+++ example/resources/static/LICENSE.EXAMPLE_RESOURCES.md
