+++ example/resources/vectorstores/LICENSE.VECTORSTORE.md
