+++ example/resources/vectorstores/byu/index.faiss
