+++ example/resources/vectorstores/maeser/index.faiss
