+++ example/tools/embeddings_example.py
